Remove "current" storage role from MMCore
It is not yet clear that having a stateful "current storage" would be
useful, especially because applications need to have knowledge of
specific storage devices to make use of them.

Let's require explicit specification of the storage label when using the
storage API (actually only applies to createDataset() at the moment).

(If ever resurrecting this, beware that handling of CoreStorage was
missing from CorePropertyCollection::Refresh().)